adding skeleton of project
this first shot includes a GCP client (what I have) and I am
up to the step where we have cached data that can be loaded
into a custom instance class. Next I am going to write a common
"rosetta stone" for attributes we are interested in, and define
those features for the instance group select function. When
that is done, I think it would be fun to do a quick clingo
solve to find an ideal set. After that I can add AWS.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>